Update Courier tests

+++ src/test/java/CreateCourierTest.java
@@ -1,12 +1,10 @@
 import io.restassured.response.ValidatableResponse;
-import org.example.courier.CourierCheck;
-import org.example.courier.CourierClient;
-import org.example.courier.CourierGenerator;
-import org.example.courier.Credentials;
+import org.example.courier.*;
 import io.qameta.allure.junit4.DisplayName;
 import io.qameta.allure.Description;
 import io.qameta.allure.Step;
 import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 
 
@@ -14,67 +12,51 @@ public class CreateCourierTest {
 
     private final CourierClient client = new CourierClient();
     private final CourierCheck check = new CourierCheck();
-    protected int courierId;
+    protected Courier courier;
+
+    @Before
+    public void createCourier() {
+        courier = CourierGenerator.random();
+    }
 
 
     @Test
     @DisplayName("Check courier creation")
     @Description("Basic courier creation test")
-    public void createCourier() {
-        var courier = CourierGenerator.random();
-
+    public void createCourierSuccessfully() {
         ValidatableResponse response = client.create(courier);
         check.createdCourierSuccessfully(response);
-
-        var creds = Credentials.from(courier);
-
-        ValidatableResponse loginResponse = client.login(creds);
-        courierId = check.loggedInSuccessfully(loginResponse);
     }
 
     @Test
     @DisplayName("Check the creation of two identical couriers")
     @Description("Check that it is impossible to create two identical couriers")
     public void createTwoIdenticalCouriers() {
-        var courier = CourierGenerator.random();
-
         ValidatableResponse response = client.create(courier);
         ValidatableResponse secondResponse = client.create(courier);
 
         check.createdCourierSuccessfully(response);
         check.createdCourierUnsuccessfully(secondResponse);
-
-        var creds = Credentials.from(courier);
-
-        ValidatableResponse loginResponse = client.login(creds);
-        courierId = check.loggedInSuccessfully(loginResponse);
     }
 
     @Test
     @DisplayName("Check the creation of two couriers with identical login")
     @Description("Check that it is impossible to create two couriers with identical login")
     public void createTwoIdenticalCouriersLogin() {
-        var firstCourier = CourierGenerator.random();
-        var secondCourier = CourierGenerator.random();
-        secondCourier.setLogin(firstCourier.getLogin());
+        var identicalLoginCourier = CourierGenerator.random();
+        identicalLoginCourier.setLogin(courier.getLogin());
 
-        ValidatableResponse response = client.create(firstCourier);
-        ValidatableResponse secondResponse = client.create(secondCourier);
+        ValidatableResponse response = client.create(courier);
+        ValidatableResponse secondResponse = client.create(identicalLoginCourier);
 
         check.createdCourierSuccessfully(response);
         check.createdCourierUnsuccessfully(secondResponse);
-
-        var creds = Credentials.from(firstCourier);
-
-        ValidatableResponse loginResponse = client.login(creds);
-        courierId = check.loggedInSuccessfully(loginResponse);
     }
 
     @Test
     @DisplayName("Check courier creation without login field")
     @Description("Check that it is impossible to create courier without login field")
     public void createCourierWithoutLogin() {
-        var courier = CourierGenerator.random();
         courier.setLogin(null);
 
         ValidatableResponse response = client.create(courier);
@@ -85,7 +67,6 @@ public class CreateCourierTest {
     @DisplayName("Check courier creation without password field")
     @Description("Check that it is impossible to create courier without password field")
     public void createCourierWithoutPassword() {
-        var courier = CourierGenerator.random();
         courier.setPassword(null);
 
         ValidatableResponse response = client.create(courier);
@@ -96,20 +77,23 @@ public class CreateCourierTest {
     @DisplayName("Check courier creation without firstName field")
     @Description("Check that it is impossible to create courier without firstName field")
     public void createCourierWithoutFirstName() {
-        var courier = CourierGenerator.random();
         courier.setFirstName(null);
 
         ValidatableResponse response = client.create(courier);
         check.createdCourierSuccessfully(response);
-
-        var creds = Credentials.from(courier);
-
-        ValidatableResponse loginResponse = client.login(creds);
-        courierId = check.loggedInSuccessfully(loginResponse);
     }
 
     @After
     public void deleteCourier() {
+        if (!courier.isCorrect()) {
+            return;
+        }
+
+        var creds = Credentials.from(courier);
+
+        ValidatableResponse loginResponse = client.login(creds);
+        var courierId = check.loggedInSuccessfully(loginResponse);
+
         if (courierId != 0) {
             ValidatableResponse delete = client.delete(courierId);
             check.deletedCourierSuccessfully(delete);
